fn-send-email/requirements.txt
functions-framework==3.8.2
google-cloud-storage==3.0.0
google-cloud-secret-manager==2.16.2
google-cloud-bigquery==3.30.0